feat(commands): add explicit RunServerCommand::register_http_routes_from_inventory
Promote the hidden `auto_register_router()` pre-dispatch hook into a
named, `pub` method on `RunServerCommand` so the HTTP route-registration
step is visible at the command's call site (Refs #4453 DP-1).

The new method iterates `UrlPatternsRegistration` inventory entries,
validates that exactly one `#[routes]` function is reachable, and
installs the resulting `ServerRouter` globally. A DP-7 confusable-API
guard rejects calls made after a manual `register_router(..)` so the
manual setter and the inventory consumer remain mutually exclusive.

`RunServerCommand::execute(..)` now invokes the method explicitly,
short-circuiting through `is_router_registered()` for opt-out users
who hand-build a router and pre-register it themselves. The legacy
`auto_register_router()` is preserved as a thin delegation so
non-`runserver` HTTP-serving commands (`showurls`, `generateopenapi`,
`introspect`) continue working until their own explicit
`register_*_from_inventory()` plumbing is introduced. `requires_router`
no longer returns true for `Runserver` to avoid double-consumption.

Refs #4453
